pyproject.toml
[project]
name = "pl-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.0",
    "detect-secrets>=1.5.0",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "httpx-oauth>=0.16.1",
    "jupyterlab>=4.4.3",
    "oauth2client>=4.1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.45.1",
    "streamlit-auth0>=1.0.5",
    "supabase>=2.15.3",
]
